requirements.txt
streamlit>=1.30.0
openai>=1.12.0
anthropic>=0.18.0
streamlit-mermaid>=0.2.0
python-dotenv>=1.0.0
requests>=2.31.0
python-pptx>=0.6.21
